ModeratorBOT/telegram-moderation-bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiofiles==23.2.1
aiohttp==3.9.1
regex==2023.12.25
apscheduler==3.10.4
structlog==23.3.0
pyyaml==6.0.1
